improve add team button on overview
