+++ frontend/src/app/Overview/Overview.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-key */
-import React, { useState } from 'react';
+import React from 'react';
 import {
   PageSection,
   TextContent,
@@ -10,18 +10,14 @@ import {
 import { CopyIcon, PlusIcon } from '@patternfly/react-icons';
 import { InfoBanner } from './InfoBanner';
 import { About } from './About';
-import { TeamForm } from '@app/Teams/TeamForm';
+import { useHistory } from 'react-router-dom';
 
 export const Overview = () => {
-  const [isOpen, setOpen] = useState<boolean>(false);
+  const history = useHistory();
 
   const handleModalToggle = () => {
-    setOpen(true)
+    history.push("/home/teams?isOpen=true")
   };
-
-  function handleChange(event, open) {
-    setOpen(open)
-  }
 
   return (
     <React.Fragment>
@@ -44,10 +40,6 @@ export const Overview = () => {
             By creating a team or joining an existing one, you can be more informed about the code coverage, OpenShift CI prow jobs, and GitHub actions of the StoneSoup components.
           </Text>
           <Button onClick={handleModalToggle} type="button" variant="primary"> <PlusIcon></PlusIcon> Add Team </Button>
-          <TeamForm
-            isOpen={isOpen}
-            handleChange={(event, open) => handleChange(event, open)}
-          ></TeamForm>
         </TextContent>
       </PageSection>
       <PageSection>
